requirements.txt
langchain
streamlit
openai
chromadb <= 0.3.29
tiktoken
pypdf
unstructured[all-docs]
pdfminer
poppler-utils
python-dotenv
azure-search-documents
azure-storage-blob
azure-identity